Development

+++ app/src/main/java/com/example/fahadali/diabetesapp/Model/BloodSugar.java
@@ -2,7 +2,6 @@ package com.example.fahadali.diabetesapp.Model;
 
 import java.sql.Date;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 
 
 public class BloodSugar {
@@ -37,7 +36,7 @@ public class BloodSugar {
         return timeString;
     }
 
-    public void setBloodSugar(int bloodSugar) {
+    public void setBloodSugar(double bloodSugar) {
         this.bloodSugar = bloodSugar;
     }
